[BE] FIX: User Delete Bug

See merge request s07-webmobile2-sub2/S07P12B103!101

+++ backend/eco2/src/main/java/com/web/eco2/model/service/user/ProfileImgService.java
@@ -82,11 +82,13 @@ public class ProfileImgService {
 
     public void deleteImage(Long userId) {
         ProfileImg img = profileImgRepository.getByUser_Id(userId);
+        System.out.println("img=====    "+img);
+        if(img.getSaveFolder() != null) {
+            Path path = Paths.get(img.getSaveFolder());
+            File realFile = new File(path.toAbsolutePath().toString(), img.getSaveName());
+            realFile.delete();
 
-        Path path = Paths.get(img.getSaveFolder());
-        File realFile = new File(path.toAbsolutePath().toString(), img.getSaveName());
-        realFile.delete();
-
+        }
         profileImgRepository.deleteById(userId);
     }
